refactor(Type): 숫자 타입 변환
- 계산기 기능을 할 때 int 타입의 오버플로우를 넘어서는 숫자를 입력할 수 있음
- 이때 오버플로우가 발생하지 않도록 충분히 큰 타입인 long 타입으로 변환

+++ src/main/java/calculator/Application.java
@@ -7,7 +7,7 @@ public class Application {
 
         CalculatorController CalculatorController = new CalculatorController();
         UserInputDto userinputDto = new UserInputDto();
-        int result;
+        long result;
 
         System.out.println("덧셈할 문자열을 입력해 주세요.");
         userinputDto.setInput(Console.readLine());
